Bug 360829: enabling new warnings and correcting code as necessary.

+++ js/tamarin/MMgc/GCMemoryProfiler.cpp
@@ -421,7 +421,7 @@ namespace MMgc
 		if(size == 0)
 			return;
 
-		if (*endMarker != 0xdeadbeef)
+		if (*endMarker != (int32)0xdeadbeef)
 		{
 			// if you get here, you have a buffer overrun.  The stack trace about to
 			// be printed tells you where the block was allocated from.  To find the
@@ -533,7 +533,7 @@ namespace MMgc
 			PrintStackTraceByIndex(*((int*)item - 1)); 
 	}
 
-	void PrintStackTraceByIndex(unsigned int i)
+	void PrintStackTraceByIndex(int i)
 	{
 #ifdef GCHEAP_LOCK
 		GCEnterCriticalSection lock(m_traceTableLock);
